add desktop

+++ setup.sh
@@ -8,12 +8,20 @@ fi
 echo "Link .spacemacs"
 ln -s $THIS_DIR/.spacemacs $HOME/.spacemacs	
 
-echo "replace ~/.emacs.d"
-rm ~/.emacs.d -rf
-cp ./spacemacs ~/.emacs.d -r
+#show replace manual
+if ! [[ -e ~/.emacs.d ]]; then
+    echo "replace ~/.emacs.d"
+    rm ~/.emacs.d -rf
+    cp ./spacemacs ~/.emacs.d -r
+fi
 
 if [[ -e $HOME/.emacs.d/private ]] ;then
     rm $HOME/.emacs.d/private -rf
-echo "Link private"
-ln -s $THIS_DIR $HOME/.emacs.d/private
+    echo "Link private"
+    ln -s $THIS_DIR $HOME/.emacs.d/private
+fi
+
+if ! [[ -e /usr/local/bin/chinese_emacs ]]; then
+    echo install chinese_emacs
+    sudo cp ./chinese_emacs /usr/local/bin/
 fi
